@@ -9,9 +9,11 @@ import android.support.v4.app.FragmentTransaction;
 import android.view.ViewGroup;
 
 import com.u9porn.data.db.entity.Category;
+import com.u9porn.ui.axgle.AxgleFragment;
+import com.u9porn.ui.images.huaban.HuaBanFragment;
 import com.u9porn.ui.images.meizitu.MeiZiTuFragment;
 import com.u9porn.ui.images.mm99.Mm99Fragment;
-import com.u9porn.ui.pav.PavFragment;
+import com.u9porn.ui.pxgav.PxgavFragment;
 import com.u9porn.ui.porn9forum.Forum9IndexFragment;
 import com.u9porn.ui.porn9forum.ForumFragment;
 import com.u9porn.ui.porn9video.index.IndexFragment;
@@ -103,14 +105,22 @@ public class BaseMainFragmentAdapter extends FragmentPagerAdapter {
                 MeiZiTuFragment meiZiTuFragment = MeiZiTuFragment.getInstance();
                 meiZiTuFragment.setCategory(category);
                 return meiZiTuFragment;
-            case Category.TYPE_PIG_AV:
-                PavFragment pigAvFragment = PavFragment.getInstance();
+            case Category.TYPE_PXG_AV:
+                PxgavFragment pigAvFragment = PxgavFragment.getInstance();
                 pigAvFragment.setCategory(category);
                 return pigAvFragment;
             case Category.TYPE_99_MM:
                 Mm99Fragment mm99Fragment = Mm99Fragment.getInstance();
                 mm99Fragment.setCategory(category);
                 return mm99Fragment;
+            case Category.TYPE_HUA_BAN:
+                HuaBanFragment huaBanFragment = HuaBanFragment.getInstance();
+                huaBanFragment.setCategory(category);
+                return huaBanFragment;
+            case Category.TYPE_AXGLE:
+                AxgleFragment axgleFragment=AxgleFragment.getInstance();
+                axgleFragment.setCategory(category);
+                return axgleFragment;
             default:
         }
         return new Fragment();
