@@ -1,16 +1,19 @@
 package com.u9porn.data;
 
+import android.graphics.Bitmap;
+
 import com.danikula.videocache.HttpProxyCacheServer;
 import com.u9porn.cookie.CookieManager;
 import com.u9porn.data.db.DbHelper;
 import com.u9porn.data.model.BaseResult;
 import com.u9porn.data.db.entity.Category;
 import com.u9porn.data.model.F9PronItem;
+import com.u9porn.data.model.HuaBan;
 import com.u9porn.data.model.MeiZiTu;
 import com.u9porn.data.model.Mm99;
 import com.u9porn.data.model.Notice;
-import com.u9porn.data.model.PavModel;
-import com.u9porn.data.model.PavVideoParserJsonResult;
+import com.u9porn.data.model.pxgav.PxgavResultWithBlockId;
+import com.u9porn.data.model.pxgav.PxgavVideoParserJsonResult;
 import com.u9porn.data.model.PinnedHeaderEntity;
 import com.u9porn.data.model.F9PornContent;
 import com.u9porn.data.model.ProxyModel;
@@ -19,6 +22,7 @@ import com.u9porn.data.model.UpdateVersion;
 import com.u9porn.data.model.User;
 import com.u9porn.data.model.VideoComment;
 import com.u9porn.data.db.entity.VideoResult;
+import com.u9porn.data.model.axgle.AxgleResponse;
 import com.u9porn.data.network.ApiHelper;
 import com.u9porn.data.prefs.PreferencesHelper;
 import com.u9porn.utils.UserHelper;
@@ -29,6 +33,8 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import io.reactivex.Observable;
+import okhttp3.ResponseBody;
+import retrofit2.Call;
 
 /**
  * @author flymegoc
@@ -194,6 +200,11 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
+    public Observable<Bitmap> porn9VideoLoginCaptcha() {
+        return mApiHelper.porn9VideoLoginCaptcha();
+    }
+
+    @Override
     public Observable<User> userLoginPorn9Video(String username, String password, String captcha) {
         return mApiHelper.userLoginPorn9Video(username, password, captcha);
     }
@@ -229,6 +240,11 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
+    public Observable<String> commonQuestions() {
+        return mApiHelper.commonQuestions();
+    }
+
+    @Override
     public Observable<BaseResult<List<MeiZiTu>>> listMeiZiTu(String tag, int page, boolean pullToRefresh) {
         return mApiHelper.listMeiZiTu(tag, page, pullToRefresh);
     }
@@ -249,18 +265,18 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
-    public Observable<List<PavModel>> loadPavListByCategory(String category, boolean pullToRefresh) {
-        return mApiHelper.loadPavListByCategory(category, pullToRefresh);
+    public Observable<PxgavResultWithBlockId> loadPxgavListByCategory(String category, boolean pullToRefresh) {
+        return mApiHelper.loadPxgavListByCategory(category, pullToRefresh);
     }
 
     @Override
-    public Observable<List<PavModel>> loadMorePavListByCategory(String category, int page, boolean pullToRefresh) {
-        return mApiHelper.loadMorePavListByCategory(category, page, pullToRefresh);
+    public Observable<PxgavResultWithBlockId> loadMorePxgavListByCategory(String category, int page, String lastBlockId, boolean pullToRefresh) {
+        return mApiHelper.loadMorePxgavListByCategory(category, page, lastBlockId, pullToRefresh);
     }
 
     @Override
-    public Observable<PavVideoParserJsonResult> loadPavVideoUrl(String url, String pId, boolean pullToRefresh) {
-        return mApiHelper.loadPavVideoUrl(url, pId, pullToRefresh);
+    public Observable<PxgavVideoParserJsonResult> loadPxgavVideoUrl(String url, String pId, boolean pullToRefresh) {
+        return mApiHelper.loadPxgavVideoUrl(url, pId, pullToRefresh);
     }
 
     @Override
@@ -414,16 +430,6 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
-    public void setNeverAskForWatchDownloadTip(boolean neverAskForWatchDownloadTip) {
-        mPreferencesHelper.setNeverAskForWatchDownloadTip(neverAskForWatchDownloadTip);
-    }
-
-    @Override
-    public boolean isNeverAskForWatchDownloadTip() {
-        return mPreferencesHelper.isNeverAskForWatchDownloadTip();
-    }
-
-    @Override
     public void setIgnoreUpdateVersionCode(int versionCode) {
         mPreferencesHelper.setIgnoreUpdateVersionCode(versionCode);
     }
@@ -464,22 +470,22 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
-    public void setMainFirstTabShow(int firstTabShow) {
+    public void setMainFirstTabShow(String firstTabShow) {
         mPreferencesHelper.setMainFirstTabShow(firstTabShow);
     }
 
     @Override
-    public int getMainFirstTabShow() {
+    public String getMainFirstTabShow() {
         return mPreferencesHelper.getMainFirstTabShow();
     }
 
     @Override
-    public void setMainSecondTabShow(int secondTabShow) {
+    public void setMainSecondTabShow(String secondTabShow) {
         mPreferencesHelper.setMainSecondTabShow(secondTabShow);
     }
 
     @Override
-    public int getMainSecondTabShow() {
+    public String getMainSecondTabShow() {
         return mPreferencesHelper.getMainSecondTabShow();
     }
 
@@ -514,6 +520,36 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
+    public boolean isShowUrlRedirectTipDialog() {
+        return mPreferencesHelper.isShowUrlRedirectTipDialog();
+    }
+
+    @Override
+    public void setShowUrlRedirectTipDialog(boolean showUrlRedirectTipDialog) {
+        mPreferencesHelper.setShowUrlRedirectTipDialog(showUrlRedirectTipDialog);
+    }
+
+    @Override
+    public void setAxgleAddress(String address) {
+        mPreferencesHelper.setAxgleAddress(address);
+    }
+
+    @Override
+    public String getAxgleAddress() {
+        return mPreferencesHelper.getAxgleAddress();
+    }
+
+    @Override
+    public boolean isFixMainNavigation() {
+        return mPreferencesHelper.isFixMainNavigation();
+    }
+
+    @Override
+    public void setFixMainNavigation(boolean fixMainNavigation) {
+        mPreferencesHelper.setFixMainNavigation(fixMainNavigation);
+    }
+
+    @Override
     public void existProxyTest() {
         mApiHelper.existProxyTest();
     }
@@ -531,6 +567,36 @@ public class AppDataManager implements DataManager {
     @Override
     public Observable<Boolean> testPavAddress(String url) {
         return mApiHelper.testPavAddress(url);
+    }
+
+    @Override
+    public Observable<Boolean> testAxgle() {
+        return mApiHelper.testAxgle();
+    }
+
+    @Override
+    public Observable<List<HuaBan.Picture>> findPictures(int categoryId, int page) {
+        return mApiHelper.findPictures(categoryId, page);
+    }
+
+    @Override
+    public Observable<AxgleResponse> axgleVideos(int page, String o, String t, String type, String c, int limit) {
+        return mApiHelper.axgleVideos(page, o, t, type, c, limit);
+    }
+
+    @Override
+    public Observable<AxgleResponse> searchAxgleVideo(String keyWord, int page) {
+        return mApiHelper.searchAxgleVideo(keyWord, page);
+    }
+
+    @Override
+    public Observable<AxgleResponse> searchAxgleJavVideo(String keyWord, int page) {
+        return mApiHelper.searchAxgleJavVideo(keyWord, page);
+    }
+
+    @Override
+    public Call<ResponseBody> getPlayVideoUrl(String url) {
+        return mApiHelper.getPlayVideoUrl(url);
     }
 
     @Override
