@@ -10,6 +10,7 @@ public interface Api {
     String APP_MEIZITU_DOMAIN = "http://www.mzitu.com/";
     String APP_99_MM_DOMAIN = "http://www.99mm.me/";
     String APP_PROXY_XICI_DAILI_DOMAIN = "http://www.xicidaili.com/";
+    String APP_HUA_BAN_DOMAIN = "http://flymegoc.cn/";
 
     String GITHUB_DOMAIN_NAME = "github";
     String PORN9_VIDEO_DOMAIN_NAME = "porn9_video_domain_name";
@@ -17,5 +18,7 @@ public interface Api {
     String PA_DOMAIN_NAME = "pv_domain_name";
     String MEI_ZI_TU_DOMAIN_NAME = "mei_zi_tu_domain_name";
     String MM_99_DOMAIN_NAME = "mm_99_domain_name";
-    String XICI_DAILI_DOMAIN_NAME="xici_daili_domain_name";
+    String XICI_DAILI_DOMAIN_NAME = "xici_daili_domain_name";
+    String HUA_BAN_DOMAIN_NAME = "huaban";
+    String AXGLE_DOMAIN_NAME = "axgle";
 }
