@@ -1,12 +1,15 @@
 package com.u9porn.data.network;
 
+import android.graphics.Bitmap;
+
 import com.u9porn.data.model.BaseResult;
 import com.u9porn.data.model.F9PronItem;
+import com.u9porn.data.model.HuaBan;
 import com.u9porn.data.model.MeiZiTu;
 import com.u9porn.data.model.Mm99;
 import com.u9porn.data.model.Notice;
-import com.u9porn.data.model.PavModel;
-import com.u9porn.data.model.PavVideoParserJsonResult;
+import com.u9porn.data.model.pxgav.PxgavResultWithBlockId;
+import com.u9porn.data.model.pxgav.PxgavVideoParserJsonResult;
 import com.u9porn.data.model.PinnedHeaderEntity;
 import com.u9porn.data.model.F9PornContent;
 import com.u9porn.data.model.ProxyModel;
@@ -15,10 +18,13 @@ import com.u9porn.data.model.UpdateVersion;
 import com.u9porn.data.model.User;
 import com.u9porn.data.model.VideoComment;
 import com.u9porn.data.db.entity.VideoResult;
+import com.u9porn.data.model.axgle.AxgleResponse;
 
 import java.util.List;
 
 import io.reactivex.Observable;
+import okhttp3.ResponseBody;
+import retrofit2.Call;
 
 /**
  * @author flymegoc
@@ -50,6 +56,8 @@ public interface ApiHelper {
 
     Observable<List<V9PornItem>> deletePorn9MyFavoriteVideo(String rvid);
 
+    Observable<Bitmap> porn9VideoLoginCaptcha();
+
     Observable<User> userLoginPorn9Video(String username, String password, String captcha);
 
     Observable<User> userRegisterPorn9Video(String username, String password1, String password2, String email, String captchaInput);
@@ -64,19 +72,21 @@ public interface ApiHelper {
 
     Observable<Notice> checkNewNotice();
 
+    Observable<String> commonQuestions();
+
     Observable<BaseResult<List<MeiZiTu>>> listMeiZiTu(String tag, int page, boolean pullToRefresh);
 
     Observable<List<String>> meiZiTuImageList(int id, boolean pullToRefresh);
 
     Observable<BaseResult<List<Mm99>>> list99Mm(String category, int page, boolean cleanCache);
 
-    Observable<List<String>> mm99ImageList(int id, String imageUrl, boolean pullToRefresh);
+    Observable<List<String>> mm99ImageList(int id, String contentUrl, boolean pullToRefresh);
 
-    Observable<List<PavModel>> loadPavListByCategory(String category, boolean pullToRefresh);
+    Observable<PxgavResultWithBlockId> loadPxgavListByCategory(String category, boolean pullToRefresh);
 
-    Observable<List<PavModel>> loadMorePavListByCategory(String category, int page, boolean pullToRefresh);
+    Observable<PxgavResultWithBlockId> loadMorePxgavListByCategory(String category, int page, String lastBlockId, boolean pullToRefresh);
 
-    Observable<PavVideoParserJsonResult> loadPavVideoUrl(String url, String pId, boolean pullToRefresh);
+    Observable<PxgavVideoParserJsonResult> loadPxgavVideoUrl(String url, String pId, boolean pullToRefresh);
 
     Observable<BaseResult<List<ProxyModel>>> loadXiCiDaiLiProxyData(int page);
 
@@ -89,4 +99,16 @@ public interface ApiHelper {
     Observable<Boolean> testPorn9ForumAddress();
 
     Observable<Boolean> testPavAddress(String url);
+
+    Observable<Boolean> testAxgle();
+
+    Observable<List<HuaBan.Picture>> findPictures(int categoryId, int page);
+
+    Observable<AxgleResponse> axgleVideos(int page, String o, String t, String type, String c, int limit);
+
+    Observable<AxgleResponse> searchAxgleVideo(String keyWord, int page);
+
+    Observable<AxgleResponse> searchAxgleJavVideo(String keyWord, int page);
+
+    Call<ResponseBody> getPlayVideoUrl(String url);
 }
