@@ -1,10 +1,7 @@
 package com.u9porn.data.network.apiservice;
 
-import com.u9porn.data.network.Api;
-
 import io.reactivex.Observable;
 import retrofit2.http.GET;
-import retrofit2.http.Headers;
 import retrofit2.http.Url;
 
 /**
@@ -19,7 +16,6 @@ public interface GitHubServiceApi {
      * @param url 链接
      * @return ob
      */
-    @Headers({"Domain-Name: " + Api.GITHUB_DOMAIN_NAME})
     @GET
     Observable<String> checkUpdate(@Url String url);
 
@@ -29,7 +25,15 @@ public interface GitHubServiceApi {
      * @param url 链接
      * @return ob
      */
-    @Headers({"Domain-Name: " + Api.GITHUB_DOMAIN_NAME})
     @GET
     Observable<String> checkNewNotice(@Url String url);
+
+    /**
+     * 常见问题
+     *
+     * @param url 链接
+     * @return ob
+     */
+    @GET
+    Observable<String> commonQuestions(@Url String url);
 }
