@@ -17,6 +17,7 @@ public interface Mm99ServiceApi {
 
     /**
      * 图库列表
+     *
      * @param url 链接
      * @return ob
      */
@@ -26,12 +27,12 @@ public interface Mm99ServiceApi {
 
     /**
      * 对应图库图片列表
-     * @param act act
-     * @param id id
+     *
+     * @param url url
      * @return ob
      */
     @Headers({"Referer: http://www.99mm.me/meitui/",
             "Domain-Name: " + Api.MM_99_DOMAIN_NAME})
-    @GET("url.php")
-    Observable<String> imageLists(@Query("act") String act, @Query("id") int id);
+    @GET
+    Observable<String> imageLists(@Url String url);
 }
