@@ -62,10 +62,6 @@ public interface PreferencesHelper {
 
     int getProxyPort();
 
-    void setNeverAskForWatchDownloadTip(boolean neverAskForWatchDownloadTip);
-
-    boolean isNeverAskForWatchDownloadTip();
-
     void setIgnoreUpdateVersionCode(int versionCode);
 
     int getIgnoreUpdateVersionCode();
@@ -82,13 +78,13 @@ public interface PreferencesHelper {
 
     int getNoticeVersionCode();
 
-    void setMainFirstTabShow(int firstTabShow);
+    void setMainFirstTabShow(String firstTabShow);
 
-    int getMainFirstTabShow();
+    String getMainFirstTabShow();
 
-    void setMainSecondTabShow(int secondTabShow);
+    void setMainSecondTabShow(String secondTabShow);
 
-    int getMainSecondTabShow();
+    String getMainSecondTabShow();
 
     void setSettingScrollViewScrollPosition(int position);
 
@@ -101,4 +97,16 @@ public interface PreferencesHelper {
     void setCustomDownloadVideoDirPath(String customDirPath);
 
     String getCustomDownloadVideoDirPath();
+
+    boolean isShowUrlRedirectTipDialog();
+
+    void setShowUrlRedirectTipDialog(boolean showUrlRedirectTipDialog);
+
+    void setAxgleAddress(String address);
+
+    String getAxgleAddress();
+
+    boolean isFixMainNavigation();
+
+    void setFixMainNavigation(boolean fixMainNavigation);
 }
