@@ -2,7 +2,6 @@ package com.u9porn.service;
 
 import android.app.Notification;
 import android.app.PendingIntent;
-import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
 import android.support.v4.app.NotificationCompat;
@@ -10,23 +9,24 @@ import android.text.format.Formatter;
 
 import com.liulishuo.filedownloader.BaseDownloadTask;
 import com.liulishuo.filedownloader.model.FileDownloadStatus;
-import com.u9porn.MyApplication;
 import com.u9porn.R;
+import com.u9porn.constants.Constants;
 import com.u9porn.data.DataManager;
 import com.u9porn.data.db.entity.V9PornItem;
-import com.u9porn.di.component.DaggerServiceComponent;
 import com.u9porn.ui.download.DownloadActivity;
 import com.u9porn.utils.DownloadManager;
-import com.u9porn.constants.Constants;
+import com.u9porn.utils.NotificationChannelHelper;
 
 import java.util.List;
 
 import javax.inject.Inject;
 
+import dagger.android.DaggerService;
+
 /**
  * @author flymegoc
  */
-public class DownloadVideoService extends Service implements DownloadManager.DownloadStatusUpdater {
+public class DownloadVideoService extends DaggerService implements DownloadManager.DownloadStatusUpdater {
 
     @Inject
     protected DataManager dataManager;
@@ -44,7 +44,6 @@ public class DownloadVideoService extends Service implements DownloadManager.Dow
     @Override
     public void onCreate() {
         super.onCreate();
-        DaggerServiceComponent.builder().applicationComponent(((MyApplication) getApplication()).getApplicationComponent()).build().inject(this);
         DownloadManager.getImpl().addUpdater(this);
     }
 
@@ -56,8 +55,10 @@ public class DownloadVideoService extends Service implements DownloadManager.Dow
 
     private void startNotification(String videoName, int progress, String fileSize, int speed) {
         int id = Constants.VIDEO_DOWNLOAD_NOTIFICATION_ID;
-        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, String.valueOf(id));
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, NotificationChannelHelper.CHANNEL_ID_FOR_DOWNLOAD);
         builder.setContentTitle("正在下载");
+        //只响铃震动一次
+        builder.setOnlyAlertOnce(true);
         builder.setSmallIcon(R.mipmap.ic_launcher_round);
         builder.setProgress(100, progress, false);
         builder.setContentText(fileSize + "--" + speed + "KB/s");
