@@ -8,22 +8,18 @@ import android.support.annotation.IntRange;
 import android.support.annotation.Nullable;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 
 import com.jaeger.library.StatusBarUtil;
 import com.sdsmdg.tastytoast.TastyToast;
-import com.u9porn.MyApplication;
 import com.u9porn.R;
-import com.u9porn.data.db.entity.V9PornItem;
-import com.u9porn.di.component.ActivityComponent;
-import com.u9porn.di.component.DaggerActivityComponent;
-import com.u9porn.di.module.ActivityModule;
-import com.u9porn.utils.PlaybackEngine;
 import com.u9porn.constants.Keys;
+import com.u9porn.data.db.entity.V9PornItem;
+import com.u9porn.utils.PlaybackEngine;
 
 import cn.bingoogolapple.swipebacklayout.BGASwipeBackHelper;
+import dagger.android.support.DaggerAppCompatActivity;
 
 /**
  * @author flymegoc
@@ -31,27 +27,18 @@ import cn.bingoogolapple.swipebacklayout.BGASwipeBackHelper;
  * @describe
  */
 
-public abstract class BaseAppCompatActivity extends AppCompatActivity implements BGASwipeBackHelper.Delegate {
+public abstract class BaseAppCompatActivity extends DaggerAppCompatActivity implements BGASwipeBackHelper.Delegate {
     private final String TAG = this.getClass().getSimpleName();
 
     protected BGASwipeBackHelper mSwipeBackHelper;
     protected boolean existActivityWithAnimation = true;
     protected Context context;
-    private ActivityComponent mActivityComponent;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         initSwipeBackFinish();
         super.onCreate(savedInstanceState);
-        mActivityComponent = DaggerActivityComponent.builder()
-                .activityModule(new ActivityModule(this))
-                .applicationComponent(((MyApplication) getApplication()).getApplicationComponent())
-                .build();
         context = this;
-    }
-
-    public ActivityComponent getActivityComponent() {
-        return mActivityComponent;
     }
 
     @Override
