@@ -2,22 +2,23 @@ package com.u9porn.ui;
 
 import android.app.Activity;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
 
+import com.qmuiteam.qmui.widget.dialog.QMUIDialog;
 import com.sdsmdg.tastytoast.TastyToast;
-import com.u9porn.MyApplication;
 import com.u9porn.R;
+import com.u9porn.constants.Keys;
 import com.u9porn.data.db.entity.Category;
 import com.u9porn.data.db.entity.V9PornItem;
-import com.u9porn.di.component.ActivityComponent;
-import com.u9porn.di.component.DaggerActivityComponent;
-import com.u9porn.di.module.ActivityModule;
+import com.u9porn.ui.download.DownloadActivity;
+import com.u9porn.ui.main.MainActivity;
+import com.u9porn.ui.porn9video.play.BasePlayVideo;
 import com.u9porn.utils.PlaybackEngine;
-import com.u9porn.constants.Keys;
+
+import dagger.android.support.DaggerFragment;
 
 /**
  * @author flymegoc
@@ -25,7 +26,7 @@ import com.u9porn.constants.Keys;
  * @describe
  */
 
-public abstract class BaseFragment extends Fragment {
+public abstract class BaseFragment extends DaggerFragment {
     private final String TAG = getClass().getSimpleName();
     private final String KEY_SAVE_DIN_STANCE_STATE_CATEGORY = "key_save_din_stance_state_category";
 
@@ -33,21 +34,12 @@ public abstract class BaseFragment extends Fragment {
     protected Activity activity;
     protected Category category;
     protected boolean mIsLoadedData;
-    private ActivityComponent mActivityComponent;
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
         this.context = getContext();
         activity = getActivity();
-        mActivityComponent = DaggerActivityComponent.builder()
-                .activityModule(new ActivityModule((AppCompatActivity) activity))
-                .applicationComponent(((MyApplication) activity.getApplication()).getApplicationComponent())
-                .build();
-    }
-
-    public ActivityComponent getActivityComponent() {
-        return mActivityComponent;
     }
 
     @Override
@@ -153,12 +145,19 @@ public abstract class BaseFragment extends Fragment {
         }
     }
 
-    protected void goToPlayVideo(V9PornItem v9PornItem, int playBackEngine) {
+    protected void goToPlayVideo(V9PornItem v9PornItem, int playBackEngine, int skipPage, int position) {
         Intent intent = PlaybackEngine.getPlaybackEngineIntent(getContext(), playBackEngine);
         intent.putExtra(Keys.KEY_INTENT_V9PORN_ITEM, v9PornItem);
-        if (activity != null) {
+        intent.putExtra(Keys.KEY_INTENT_CATEGORY_ITEM, category);
+        intent.putExtra(Keys.KEY_INTENT_SKIP_PAGE, skipPage);
+        intent.putExtra(Keys.KEY_INTENT_SCROLL_TO_POSITION, position);
+        if (activity instanceof MainActivity || activity instanceof DownloadActivity) {
             startActivity(intent);
             activity.overridePendingTransition(R.anim.slide_in_right, R.anim.side_out_left);
+        } else if (activity instanceof BasePlayVideo) {
+            BasePlayVideo basePlayVideo = (BasePlayVideo) activity;
+            basePlayVideo.setV9PornItems(v9PornItem);
+            basePlayVideo.initData();
         } else {
             showMessage("无法获取宿主Activity", TastyToast.INFO);
         }
@@ -175,5 +174,31 @@ public abstract class BaseFragment extends Fragment {
 
     protected void showMessage(String msg, int type) {
         TastyToast.makeText(context.getApplicationContext(), msg, TastyToast.LENGTH_SHORT, type).show();
+    }
+
+    /**
+     * 展示一个可以选择的dialog
+     *
+     * @param msg    title
+     * @param checks 可以选择的条目
+     * @param check  回调
+     */
+    protected void showDialog(String msg, String[] checks, @NonNull final DialogCheck check) {
+        final QMUIDialog.CheckableDialogBuilder builder = new QMUIDialog.CheckableDialogBuilder(this.getContext());
+        QMUIDialog dialog;
+        builder.setTitle(msg);
+        builder.addItems(checks, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                check.onCheck(builder.getCheckedIndex());
+                dialog.dismiss();
+            }
+        });
+        dialog = builder.create();
+        dialog.show();
+    }
+
+    protected interface DialogCheck {
+        void onCheck(int index);
     }
 }
