@@ -3,14 +3,22 @@ package com.u9porn.ui.about;
 import android.annotation.SuppressLint;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.Environment;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+import android.widget.ImageView;
+import android.widget.RadioGroup;
 import android.widget.TextView;
 
+import com.bumptech.glide.request.target.SimpleTarget;
+import com.bumptech.glide.request.transition.Transition;
+import com.orhanobut.logger.Logger;
 import com.qmuiteam.qmui.util.QMUIPackageHelper;
 import com.qmuiteam.qmui.widget.QMUILoadingView;
 import com.qmuiteam.qmui.widget.dialog.QMUIDialog;
@@ -25,8 +33,10 @@ import com.u9porn.ui.MvpActivity;
 import com.u9porn.utils.ApkVersionUtils;
 import com.u9porn.utils.AppCacheUtils;
 import com.u9porn.utils.DialogUtils;
+import com.u9porn.utils.GlideApp;
 
 import java.io.File;
+import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
@@ -36,6 +46,8 @@ import javax.inject.Inject;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import de.greenrobot.common.io.FileUtils;
+import ru.noties.markwon.Markwon;
 
 /**
  * @author flymegoc
@@ -60,6 +72,7 @@ public class AboutActivity extends MvpActivity<AboutView, AboutPresenter> implem
 
     @Inject
     protected AboutPresenter aboutPresenter;
+    private TextView commonQuestionTextView;
 
     @SuppressLint("SetTextI18n")
     @Override
@@ -109,10 +122,22 @@ public class AboutActivity extends MvpActivity<AboutView, AboutPresenter> implem
                         startActivity(intent);
                     }
                 })
+                .addItemView(mAboutGroupListView.createItemView("赞赏我，请我喝杯咖啡"), new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        showAppreciateDialog();
+                    }
+                })
                 .addItemView(cleanCacheQMUICommonListItemView, new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
                         showChoiceCacheCleanDialog();
+                    }
+                })
+                .addItemView(mAboutGroupListView.createItemView("常见问题"), new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        showCommonQuestionsDialog();
                     }
                 })
                 .addItemView(mAboutGroupListView.createItemView(getResources().getString(R.string.about_check_update)), new View.OnClickListener() {
@@ -129,6 +154,84 @@ public class AboutActivity extends MvpActivity<AboutView, AboutPresenter> implem
                 })
                 .addTo(mAboutGroupListView);
 
+    }
+
+    private void showCommonQuestionsDialog() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this, R.style.MyDialogTheme);
+        builder.setTitle("常见问题");
+        View view = View.inflate(this, R.layout.layout_common_questions, null);
+        commonQuestionTextView = view.findViewById(R.id.tv_common_question);
+        Markwon.setMarkdown(commonQuestionTextView, "**加载中...**");
+        builder.setView(view);
+        builder.setCancelable(false);
+        builder.setPositiveButton("知道了", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                commonQuestionTextView=null;
+                dialog.dismiss();
+            }
+        });
+        builder.show();
+        presenter.commonQuestions();
+    }
+
+    private void showAppreciateDialog() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this, R.style.MyDialogTheme);
+        builder.setTitle("赞赏作者");
+        View view = View.inflate(this, R.layout.layout_appreciate_qrcode, null);
+        final ImageView imageViewWebChat = view.findViewById(R.id.iv_appreciate_qr_code_web_chat);
+        final ImageView imageViewAliPay = view.findViewById(R.id.iv_appreciate_qr_code_ali_pay);
+        final RadioGroup radioGroup = view.findViewById(R.id.rg_pay);
+        radioGroup.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(RadioGroup group, int checkedId) {
+                switch (checkedId) {
+                    case R.id.rb_web_chat:
+                        imageViewWebChat.setVisibility(View.VISIBLE);
+                        imageViewAliPay.setVisibility(View.GONE);
+                        break;
+                    case R.id.rb_ali_pay:
+                        imageViewWebChat.setVisibility(View.GONE);
+                        imageViewAliPay.setVisibility(View.VISIBLE);
+                        break;
+                }
+            }
+        });
+        builder.setView(view);
+        builder.setNegativeButton("算了，囊中羞涩", null);
+        builder.setPositiveButton("保存至相册", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                if (radioGroup.getCheckedRadioButtonId() == R.id.rb_ali_pay) {
+                    saveToSystemGallery("zhi_fu_bao", R.drawable.alipay1547141972480);
+                } else {
+                    saveToSystemGallery("wei_xin", R.drawable.mm_reward_qrcode_1547141812376);
+                }
+            }
+        });
+        builder.show();
+    }
+
+    private void saveToSystemGallery(final String name, int id) {
+        GlideApp.with(this).downloadOnly().load(id).into(new SimpleTarget<File>() {
+            @Override
+            public void onResourceReady(@NonNull File resource, @Nullable Transition<? super File> transition) {
+                File file = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES), name + ".jpg");
+                try {
+                    FileUtils.copyFile(resource, file);
+                    showMessage("保存图片成功了", TastyToast.SUCCESS);
+                    notifySystemGallery(file);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    showMessage("保存图片失败了", TastyToast.ERROR);
+                }
+            }
+        });
+    }
+
+    private void notifySystemGallery(File file) {
+
+        MediaScannerConnection.scanFile(this, new String[]{file.getAbsolutePath()}, new String[]{"image/jpeg"}, null);
     }
 
     private void showChoiceCacheCleanDialog() {
@@ -198,7 +301,6 @@ public class AboutActivity extends MvpActivity<AboutView, AboutPresenter> implem
     @NonNull
     @Override
     public AboutPresenter createPresenter() {
-        getActivityComponent().inject(this);
         return aboutPresenter;
     }
 
@@ -288,6 +390,19 @@ public class AboutActivity extends MvpActivity<AboutView, AboutPresenter> implem
     @Override
     public void countCacheFileSizeError(String message) {
         showMessage(message, TastyToast.ERROR);
+    }
+
+    @Override
+    public void loadCommonQuestionsSuccess(String mdString) {
+        Logger.t(TAG).d(mdString);
+        if (commonQuestionTextView != null) {
+            Markwon.setMarkdown(commonQuestionTextView, mdString);
+        }
+    }
+
+    @Override
+    public void loadCommonQuestionsFailure(String errorMessage, int code) {
+        showError(errorMessage);
     }
 
     private void dismissDialog() {
