@@ -2,7 +2,6 @@ package com.u9porn.ui.basemain;
 
 
 import android.content.Context;
-import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -25,10 +24,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.PopupWindow;
 
-import com.bugsnag.android.Bugsnag;
-import com.bugsnag.android.Severity;
 import com.chad.library.adapter.base.BaseViewHolder;
-import com.devbrackets.android.exomedia.util.ResourceUtil;
 import com.orhanobut.logger.Logger;
 import com.sdsmdg.tastytoast.TastyToast;
 import com.u9porn.BuildConfig;
@@ -38,7 +34,6 @@ import com.u9porn.adapter.SortCategoryAdapter;
 import com.u9porn.data.db.entity.Category;
 import com.u9porn.eventbus.LowMemoryEvent;
 import com.u9porn.ui.MvpFragment;
-import com.u9porn.ui.porn9video.Main9PronVideoFragment;
 import com.u9porn.utils.AnimationUtils;
 import com.u9porn.utils.FragmentUtils;
 
@@ -63,8 +58,7 @@ import butterknife.Unbinder;
  */
 public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMainPresenter> implements BaseMainView, View.OnClickListener, SortCategoryAdapter.OnStartDragListener {
 
-
-    private static final String TAG = Main9PronVideoFragment.class.getSimpleName();
+    private static final String TAG = BaseMainFragment.class.getSimpleName();
     @BindView(R.id.tabLayout)
     TabLayout tabLayout;
     @BindView(R.id.iv_sort_category)
@@ -118,7 +112,6 @@ public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMai
     @NonNull
     @Override
     public BaseMainPresenter createPresenter() {
-        getActivityComponent().inject(this);
         return baseMainPresenter;
     }
 
@@ -126,8 +119,8 @@ public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMai
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         unbinder = ButterKnife.bind(this, view);
-        Drawable dropDownDrawable = ResourceUtil.tintList(context, R.drawable.ic_arrow_drop_down_black_24dp, R.color.common_always_white_text_color);
-        ivSortCategory.setImageDrawable(dropDownDrawable);
+//        Drawable dropDownDrawable = ResourceUtil.tintList(context, R.drawable.ic_arrow_drop_down_black_24dp, R.color.common_always_white_text_color);
+//        ivSortCategory.setImageDrawable(dropDownDrawable);
         ivSortCategory.setOnClickListener(this);
         mBaseMainFragmentAdapter.setDestroy(isNeedDestroy());
         viewPager.setAdapter(mBaseMainFragmentAdapter);
@@ -236,7 +229,6 @@ public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMai
             presenter.updateCategoryData(sortCategoryList);
             presenter.loadCategoryData(getCategoryType());
         }
-
         isNeedInterruptOnBackPressed = false;
     }
 
@@ -396,7 +388,7 @@ public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMai
             return;
         }
         if (!BuildConfig.DEBUG) {
-            Bugsnag.notify(new Throwable(TAG + ":LowMemory,try to release some memory now!"), Severity.INFO);
+            //Bugsnag.notify(new Throwable(TAG + ":LowMemory,try to release some memory now!"), Severity.INFO);
         }
         try {
             Logger.t(TAG).d("start try to release memory ....");
@@ -422,7 +414,7 @@ public abstract class BaseMainFragment extends MvpFragment<BaseMainView, BaseMai
         } catch (Exception e) {
             e.printStackTrace();
             if (!BuildConfig.DEBUG) {
-                Bugsnag.notify(new Throwable(TAG + " tryToReleaseMemory error::", e), Severity.WARNING);
+                //Bugsnag.notify(new Throwable(TAG + " tryToReleaseMemory error::", e), Severity.WARNING);
             }
         }
     }
