@@ -1,8 +1,7 @@
 package com.u9porn.ui.porn9video.user;
 
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.net.Uri;
+import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AlertDialog;
@@ -16,21 +15,21 @@ import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.ImageView;
 
-import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions;
-import com.orhanobut.logger.Logger;
+import com.devbrackets.android.exomedia.util.ResourceUtil;
 import com.qmuiteam.qmui.util.QMUIKeyboardHelper;
+import com.qmuiteam.qmui.widget.dialog.QMUIDialog;
+import com.qmuiteam.qmui.widget.dialog.QMUIDialogAction;
 import com.sdsmdg.tastytoast.TastyToast;
 import com.u9porn.R;
+import com.u9porn.constants.Keys;
+import com.u9porn.constants.KeysActivityRequestResultCode;
+import com.u9porn.data.DataManager;
 import com.u9porn.data.model.User;
 import com.u9porn.ui.MvpActivity;
 import com.u9porn.ui.porn9video.favorite.FavoriteActivity;
 import com.u9porn.ui.porn9video.search.SearchActivity;
 import com.u9porn.ui.setting.SettingActivity;
-import com.u9porn.utils.AddressHelper;
 import com.u9porn.utils.DialogUtils;
-import com.u9porn.utils.GlideApp;
-import com.u9porn.constants.Keys;
 
 import javax.inject.Inject;
 
@@ -41,9 +40,6 @@ import butterknife.ButterKnife;
  * @author flymegoc
  */
 public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> implements UserView {
-
-    public static final int LOGIN_ACTION_FOR_LOOK_MY_FAVORITE = 1;
-    public static final int LOGIN_ACTION_FOR_SEARCH_91PRON_VIDEO = 2;
 
     private static final String TAG = UserLoginActivity.class.getSimpleName();
     @BindView(R.id.et_account)
@@ -60,19 +56,18 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
     Toolbar toolbar;
     @BindView(R.id.cb_remember_password)
     CheckBox cbRememberPassword;
-    @BindView(R.id.cb_auto_login)
-    CheckBox cbAutoLogin;
 
     private AlertDialog alertDialog;
     private String username;
     private String password;
     private int loginForAction;
 
-    @Inject
-    protected AddressHelper addressHelper;
 
     @Inject
     UserPresenter userPresenter;
+
+    @Inject
+    DataManager dataManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -80,8 +75,9 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
         setContentView(R.layout.activity_user_login);
         ButterKnife.bind(this);
         initToolBar(toolbar);
+
         loginForAction = getIntent().getIntExtra(Keys.KEY_INTENT_LOGIN_FOR_ACTION, 0);
-        if (!TextUtils.isEmpty(addressHelper.getVideo9PornAddress())) {
+        if (!TextUtils.isEmpty(presenter.getVideo9PornAddress())) {
             loadCaptcha();
         }
         btUserLogin.setOnClickListener(new View.OnClickListener() {
@@ -100,17 +96,17 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
             }
         });
 
-        cbAutoLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (cbAutoLogin.isChecked()) {
-                    cbAutoLogin.setChecked(true);
-                    cbRememberPassword.setChecked(true);
-                } else {
-                    cbAutoLogin.setChecked(false);
-                }
-            }
-        });
+//        cbAutoLogin.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View v) {
+//                if (cbAutoLogin.isChecked()) {
+//                    cbAutoLogin.setChecked(true);
+//                    cbRememberPassword.setChecked(true);
+//                } else {
+//                    cbAutoLogin.setChecked(false);
+//                }
+//            }
+//        });
 
         cbRememberPassword.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -119,13 +115,17 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
                     cbRememberPassword.setChecked(true);
                 } else {
                     cbRememberPassword.setChecked(false);
-                    cbAutoLogin.setChecked(false);
+                    // cbAutoLogin.setChecked(false);
                 }
             }
         });
 
         alertDialog = DialogUtils.initLoadingDialog(this, "登录中，请稍后...");
         setUpUserInfo();
+
+        if (TextUtils.isEmpty(presenter.getVideo9PornAddress())) {
+            showNeedSetAddressFirstDialog();
+        }
     }
 
     private void setUpUserInfo() {
@@ -134,8 +134,8 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
         if (!TextUtils.isEmpty(password)) {
             cbRememberPassword.setChecked(true);
         }
-        boolean isAutoLogin = presenter.isAutoLogin();
-        cbAutoLogin.setChecked(isAutoLogin);
+        //boolean isAutoLogin = presenter.isAutoLogin();
+        //cbAutoLogin.setChecked(isAutoLogin);
 
         etAccount.setText(username);
         etPassword.setText(password);
@@ -163,21 +163,12 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
      * 加载验证码，目前似乎是非必须，不填也是可以登录的
      */
     private void loadCaptcha() {
-        String url = addressHelper.getVideo9PornAddress() + "captcha.php";
-
-        Logger.t(TAG).d("验证码链接：" + url);
-        Uri uri = Uri.parse(url);
-        GlideApp.with(this).load(uri).placeholder(R.drawable.placeholder).transition(new DrawableTransitionOptions().crossFade(300)).diskCacheStrategy(DiskCacheStrategy.NONE).skipMemoryCache(true).into(captchaImageView);
+        presenter.loadCaptcha();
     }
 
     @NonNull
     @Override
     public UserPresenter createPresenter() {
-        getActivityComponent().inject(this);
-
-        if (TextUtils.isEmpty(addressHelper.getVideo9PornAddress())) {
-            showNeedSetAddressFirstDialog();
-        }
         return userPresenter;
     }
 
@@ -185,20 +176,14 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
         AlertDialog.Builder builder = new AlertDialog.Builder(this, R.style.MyDialogTheme);
         builder.setTitle("温馨提示");
         builder.setMessage("还未设置91porn视频地址,无法登录或注册，现在去设置？");
-        builder.setPositiveButton("去设置", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                Intent intent = new Intent(context, SettingActivity.class);
-                startActivityWithAnimation(intent);
-                finish();
-            }
+        builder.setPositiveButton("去设置", (dialog, which) -> {
+            Intent intent = new Intent(context, SettingActivity.class);
+            startActivityWithAnimation(intent);
+            finish();
         });
-        builder.setNegativeButton("退出", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                dialog.dismiss();
-                onBackPressed();
-            }
+        builder.setNegativeButton("退出", (dialog, which) -> {
+            dialog.dismiss();
+            onBackPressed();
         });
         builder.setCancelable(false);
         builder.show();
@@ -214,7 +199,7 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
     @Override
     public void loginSuccess(User user) {
 
-        presenter.saveUserInfoPrf(username, password, cbRememberPassword.isChecked(), cbAutoLogin.isChecked());
+        presenter.saveUserInfoPrf(username, password, cbRememberPassword.isChecked(), false);
         showMessage("登录成功", TastyToast.SUCCESS);
         switchWhereToGo();
     }
@@ -222,15 +207,23 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
 
     private void switchWhereToGo() {
         switch (loginForAction) {
-            case LOGIN_ACTION_FOR_LOOK_MY_FAVORITE:
+            case KeysActivityRequestResultCode.LOGIN_ACTION_FOR_LOOK_MY_FAVORITE:
                 Intent intent = new Intent(this, FavoriteActivity.class);
                 startActivityWithAnimation(intent);
                 finish();
                 break;
-            case LOGIN_ACTION_FOR_SEARCH_91PRON_VIDEO:
+            case KeysActivityRequestResultCode.LOGIN_ACTION_FOR_SEARCH_91PRON_VIDEO:
                 Intent intentSearch = new Intent(this, SearchActivity.class);
                 startActivityWithAnimation(intentSearch);
                 finish();
+                break;
+            case KeysActivityRequestResultCode.LOGIN_ACTION_FOR_GET_UID:
+                setResult(KeysActivityRequestResultCode.RESULT_CODE_FOR_REFRESH_GET_UID);
+                onBackPressed();
+                break;
+            case KeysActivityRequestResultCode.LOGIN_ACTION_FOR_LOOK_AUTHOR_VIDEO:
+                setResult(KeysActivityRequestResultCode.RESULT_FOR_LOOK_AUTHOR_VIDEO);
+                onBackPressed();
                 break;
             default:
                 setResult(RESULT_OK);
@@ -254,8 +247,19 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
     }
 
     @Override
-    public void showError(String message) {
+    public void loadCaptchaSuccess(Bitmap bitmap) {
+        captchaImageView.setImageBitmap(bitmap);
+    }
 
+    @Override
+    public void loadCaptchaFailure(String errorMessage, int code) {
+        captchaImageView.setImageDrawable(ResourceUtil.getDrawable(this, R.drawable.ic_refresh));
+        showError("无法加载验证码,点击刷新重试");
+    }
+
+    @Override
+    public void showError(String message) {
+        showMessage(message, TastyToast.ERROR);
     }
 
     @Override
@@ -295,8 +299,15 @@ public class UserLoginActivity extends MvpActivity<UserView, UserPresenter> impl
 
         //noinspection SimplifiableIfStatement
         if (id == R.id.action_user_register) {
-            Intent intent = new Intent(this, UserRegisterActivity.class);
-            startActivityWithAnimation(intent);
+            new QMUIDialog.MessageDialogBuilder(this)
+                    .setMessage("注册功能已停止支持，请去9*porn官网注册，之后再来登录！")
+                    .addAction("知道了", new QMUIDialogAction.ActionListener() {
+                        @Override
+                        public void onClick(QMUIDialog dialog, int index) {
+                            dialog.dismiss();
+                        }
+                    })
+                    .show();
             return true;
         }
 
