@@ -1,6 +1,7 @@
 package com.u9porn.ui.porn9video.user;
 
 import android.content.Intent;
+import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -21,7 +22,6 @@ import com.u9porn.R;
 import com.u9porn.data.model.User;
 import com.u9porn.ui.MvpActivity;
 import com.u9porn.ui.main.MainActivity;
-import com.u9porn.utils.AddressHelper;
 import com.u9porn.utils.DialogUtils;
 import com.u9porn.utils.GlideApp;
 
@@ -55,9 +55,6 @@ public class UserRegisterActivity extends MvpActivity<UserView, UserPresenter> i
     private AlertDialog alertDialog;
     private String username;
     private String password;
-
-    @Inject
-    protected AddressHelper addressHelper;
 
     @Inject
     protected UserPresenter userPresenter;
@@ -137,7 +134,6 @@ public class UserRegisterActivity extends MvpActivity<UserView, UserPresenter> i
     @NonNull
     @Override
     public UserPresenter createPresenter() {
-        getActivityComponent().inject(this);
         return userPresenter;
     }
 
@@ -145,7 +141,7 @@ public class UserRegisterActivity extends MvpActivity<UserView, UserPresenter> i
      * 加载验证码，目前似乎是非必须，不填也是可以登录的
      */
     private void loadCaptcha() {
-        String url = addressHelper.getVideo9PornAddress() + "captcha2.php";
+        String url = presenter.getVideo9PornAddress() + "captcha2.php";
         Logger.t(TAG).d("验证码链接：" + url);
         Uri uri = Uri.parse(url);
         GlideApp.with(this).load(uri).placeholder(R.drawable.placeholder).transition(new DrawableTransitionOptions().crossFade(300)).diskCacheStrategy(DiskCacheStrategy.NONE).skipMemoryCache(true).into(wbCaptcha);
@@ -171,6 +167,16 @@ public class UserRegisterActivity extends MvpActivity<UserView, UserPresenter> i
     @Override
     public void registerFailure(String message) {
         showMessage(message, TastyToast.ERROR);
+    }
+
+    @Override
+    public void loadCaptchaSuccess(Bitmap bitmap) {
+
+    }
+
+    @Override
+    public void loadCaptchaFailure(String errorMessage, int code) {
+
     }
 
     @Override
