@@ -1,6 +1,6 @@
 package com.u9porn.ui.splash;
 
-import com.u9porn.ui.porn9video.user.IBaseUser;
+import com.u9porn.ui.notice.IBaseNotice;
 
 /**
  *
@@ -8,12 +8,11 @@ import com.u9porn.ui.porn9video.user.IBaseUser;
  * @date 2017/12/21
  */
 
-public interface ISplash extends IBaseUser{
+public interface ISplash extends IBaseNotice {
+
+    int getIgnoreUpdateVersionCode();
+
     boolean isUserLogin();
 
-    String getPorn9VideoLoginUserName();
-
-    String getPorn9VideoLoginUserPassword();
-
-    boolean isPorn9VideoUserAutoLogin();
+    String getVideo9PornAddress();
 }
