@@ -1,15 +1,12 @@
 package com.u9porn.ui.splash;
 
-import com.u9porn.data.model.User;
-import com.u9porn.ui.BaseView;
+import com.u9porn.ui.notice.NoticeView;
 
 /**
  * @author flymegoc
  * @date 2017/12/21
  */
 
-public interface SplashView extends BaseView{
-    void loginSuccess(User user);
+public interface SplashView extends NoticeView {
 
-    void loginError(String message);
 }
