@@ -8,7 +8,7 @@ import android.os.Environment;
  */
 
 public class SDCardUtils {
-    private static final String ROOT_FOLDER = Environment.getExternalStorageDirectory() + "/9porn/";
+    private static final String ROOT_FOLDER = Environment.getExternalStorageDirectory() + "/v9porn/";
     public static final String DOWNLOAD_VIDEO_PATH = ROOT_FOLDER + "video/";
     public static final String DOWNLOAD_IMAGE_PATH = ROOT_FOLDER + "image/";
     public static final String DATE_FORMAT_YEAR_MONTH_DAY_HOUR_MINUTE_SECOND = "yyyy-MM-dd HH:mm:ss";
